requirements.txt
beautifulsoup4==4.8.2
celery==4.4.0
cloudscraper==1.2.20
Django==2.2.10
django-chartjs==2.0.0
django-cool-pagination==0.3.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-relativedelta==1.1.0
django-weasyprint==0.5.5
djangorestframework==3.11.0
email-validator==1.0.5
gunicorn==20.0.4
iso3166==1.0.1
Jinja2==2.11.1
jsonfield==2.0.2
pymongo==3.7.2
PySocks==1.7.1
Pillow==6.1.0
requests==2.21.0
reportlab==3.5.34
selenium==3.141.0
sqlparse==0.2.4
tbselenium==0.4.2
validators==0.14.2
